% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{directory_exists}
\alias{directory_exists}
\title{directory_exists}
\usage{
directory_exists(directory, create = "FALSE")
}
\arguments{
\item{directory, }{the directory to check}

\item{create, }{directory if true}
}
\description{
Return directory name if exists. If create is TRUE, creates directory, else stop
}
