% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_download_file}
\alias{gx_download_file}
\title{gx_download_file}
\usage{
gx_download_file(encoded_dataset_id, file_path, force)
}
\arguments{
\item{encoded_dataset_id, }{the data's encoded IDs}

\item{file_path, }{path to download to}

\item{force, }{force the download?}
}
\description{
Download a file from
}
