% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_get_collection}
\alias{gx_get_collection}
\title{gx_get_collection}
\usage{
gx_get_collection(file_id, hist_datasets, create = FALSE, force = FALSE)
}
\arguments{
\item{file_id, }{ID number}

\item{hist_datasets, }{table of all dataset info}

\item{create, }{if TRUE, create import directory if it does not exist}

\item{force, }{if TRUE, will download the file even if it already exists locally}
}
\description{
Correctly download the collection data into directories similar to gx_get()
gx_get is the parent call, in gx_get is where file_id is detemined to be a collection or not
If the file_id is a collection then this function will be called.
}
