% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_latest_history}
\alias{gx_latest_history}
\title{gx_latest_history}
\usage{
gx_latest_history()
}
\description{
Uses the Galaxy API histories/most_recently_used to get the last updated history
}
