% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_put}
\alias{gx_put}
\title{gx_put}
\usage{
gx_put(filepath, filename = "", file_type = "auto")
}
\arguments{
\item{filepath, }{Path to file}

\item{filename, }{Name of the file to display in Galaxy}

\item{file_type, }{auto-detect otherwise}
}
\description{
This function uploads a dataset to the current Galaxy history
}
