% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_restore}
\alias{gx_restore}
\title{gx_restore}
\usage{
gx_restore(rdata_id, rhistory_id)
}
\arguments{
\item{rdata_id, }{.RData ID number}

\item{rhistory_id, }{.RHistory ID number}
}
\description{
Restore the notebook from a .RData and .RHistory object from the current Galaxy history.
}
