% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_set_import_directory}
\alias{gx_set_import_directory}
\title{gx_set_import_directory}
\usage{
gx_set_import_directory(IMPORT_DIRECTORY = NULL, create = FALSE)
}
\arguments{
\item{IMPORT_DIRECTORY,}{path to use as import directory, default /tmp/<username>/galaxy_import}

\item{create,}{default FALSE. If TRUE, try to create the directory if it doesn't exist}
}
\description{
This function sets the import directory
}

