% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_set_tmp_directory}
\alias{gx_set_tmp_directory}
\title{gx_set_tmp_directory}
\usage{
gx_set_tmp_directory(TMP_DIRECTORY = NULL, create = FALSE)
}
\arguments{
\item{TMP_DIRECTORY,}{path to use as tmp directory, default $GX_IMPORT_DIRECTORY/tmp}

\item{create,}{default FALSE. If TRUE, try to create the directory if it doesn't exist}
}
\description{
This function sets the tmp directory
}

