% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_show_dataset}
\alias{gx_show_dataset}
\title{gx_show_dataset}
\usage{
gx_show_dataset(dataset_encoded_id)
}
\arguments{
\item{dataset_encoded_id, }{the encoded dataset id which you can get from gx_list_history_datasets()}
}
\description{
Show dataset info based
}
