% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_switch_history}
\alias{gx_switch_history}
\title{gx_switch_history}
\usage{
gx_switch_history(HISTORY_ID)
}
\arguments{
\item{HISTORY_ID, }{the Galaxy history id to work on}
}
\description{
Convenience method to set the current history id in the environment setting
}
