% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\name{gx_verify_collection}
\alias{gx_verify_collection}
\title{gx_verify_collection}
\usage{
gx_verify_collection(file_id, hist_datasets)
}
\arguments{
\item{file_id, }{ID number}

\item{hist_datasets, }{Datasets from Galaxy history}
}
\description{
gx_verify_collection
}
