% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{pkg.env
private environment}
\format{
An object of class \code{environment} of length 5.
}
\usage{
pkg.env
}
\description{
pkg.env
private environment
}
\keyword{datasets}
