% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GalaxyConnector.R
\docType{data}
\name{pkg.env}
\alias{pkg.env}
\title{pkg.env
private environment}
\format{\preformatted{<environment: 0x7fe28e8c7f08> 
}}
\usage{
pkg.env
}
\description{
pkg.env
private environment
}
\keyword{datasets}

